refactor: moved all pages to their own spots
