+++ src/pages/contact/contact.tsx
@@ -1,8 +1,9 @@
 import React from "react";
+import './styles.css'
 
 export function Contact() {
 
-    return(<div>
+    return(<div className="contact">
         <h1>Contact</h1>
     </div>)
 }
