+++ src/pages/gallery/gallery.tsx
@@ -1,9 +1,10 @@
 import React from "react";
+import './styles.css'
 
 
 export function Gallery() {
     
-        return(<div>
+        return(<div className="gallery">
             <h1>Gallery</h1>
         </div>)
     }
