+++ src/pages/home/Home.tsx
@@ -1,0 +1,15 @@
+import { useState } from 'react'
+
+import './styles.css'
+import { Header } from '../../modules'
+
+export function Home() {
+  const [count, setCount] = useState(0)
+
+  return (
+    <div className="home">
+    <h1>Home</h1>
+      
+    </div>
+  )
+}
