+++ src/pages/index.tsx
@@ -1,5 +1,6 @@
-import { Contact } from "./contact";
-import { Gallery } from "./gallery";
-import { Faq } from "./faq";
+import { Contact } from "./contact/contact";
+import { Gallery } from "./gallery/gallery";
+import { Faq } from "./faq/faq";
+import { Home } from "./home/Home";
 
-export{ Contact, Gallery, Faq};
+export{ Contact, Gallery, Faq, Home};
